+++ petab/parameters.py
@@ -4,7 +4,7 @@ import numbers
 import pandas as pd
 import numpy as np
 from collections import OrderedDict
-from typing import Iterable, Set, List, Tuple
+from typing import Iterable, Set, List, Tuple, Dict
 
 import libsbml
 
@@ -49,15 +49,32 @@ def write_parameter_df(df: pd.DataFrame, filename: str) -> None:
 
 def get_optimization_parameters(parameter_df: pd.DataFrame) -> List[str]:
     """
-    Get list of optimization parameter ids from parameter dataframe.
+    Get list of optimization parameter IDs from parameter table.
 
     Arguments:
         parameter_df: PEtab parameter DataFrame
 
     Returns:
-        List of parameter IDs in the parameter table
+        List of IDs of parameters selected for optimization.
     """
-    return list(parameter_df.reset_index()[PARAMETER_ID])
+    return list(parameter_df.index[parameter_df[ESTIMATE] == 1])
+
+
+def get_optimization_parameter_scaling(
+        parameter_df: pd.DataFrame) -> Dict[str, str]:
+    """
+    Get Dictionary with optimization parameter IDs mapped to parameter scaling
+    strings.
+
+    Arguments:
+        parameter_df: PEtab parameter DataFrame
+
+    Returns:
+        Dictionary with optimization parameter IDs mapped to parameter scaling
+        strings.
+    """
+    estimated_df = parameter_df.loc[parameter_df[ESTIMATE] == 1]
+    return dict(zip(estimated_df.index, estimated_df[PARAMETER_SCALE]))
 
 
 def create_parameter_df(sbml_model: libsbml.Model,
@@ -204,13 +221,13 @@ def get_valid_parameters_for_parameter_table(
     # {observable,noise}Parameters
     placeholders = set()
     for k, v in observables.items():
-        placeholders |= measurements.get_placeholders(
+        placeholders |= set(measurements.get_placeholders(
             v['formula'],
             core.get_observable_id(k),
-            'observable')
+            'observable'))
     for k, v in sigmas.items():
-        placeholders |= measurements.get_placeholders(
-            v, core.get_observable_id(k), 'noise')
+        placeholders |= set(measurements.get_placeholders(
+            v, core.get_observable_id(k), 'noise'))
 
     # exclude rule targets
     assignment_targets = {ar.getVariable()
@@ -287,23 +304,6 @@ def get_priors_from_df(parameter_df: pd.DataFrame,
         prior_list.append((prior_type, prior_pars, par_scale, par_bounds))
 
     return prior_list
-
-
-def parameter_id_is_valid(parameter_id: str) -> bool:
-    """Check whether parameter_id is a valid PEtab parameter ID
-
-    This should pretty much correspond to what is allowed for SBML identifiers.
-
-    TODO(#179) improve checking
-
-    Arguments:
-        parameter_id: Parameter ID to validate
-
-    Returns:
-        ``True`` if valid, ``False`` otherwise
-    """
-
-    return parameter_id != ''
 
 
 def scale(parameter: numbers.Number, scale_str: 'str') -> numbers.Number: